Change post schema, adjust swagger accordingly. fix tests to match new scheme and further improvements.

+++ tests/post.test.ts
@@ -4,64 +4,60 @@ import mongoose from "mongoose";
 import Post from "../models/post_model.js";
 
 const user = {
-  username: "user",
-  email: "user@email.com",
+  identifier: "test",
   password: "Password123",
 };
-let token: string;
+let tokens: { accessToken: string; refreshToken: string; userId: string };
 
 beforeAll(async () => {
-  await Post.deleteMany();
-  await request(app).post("/auth/register").send(user);
-  token = (
-    await request(app).post("/auth/login").send({
-      identifier: user.username,
-      password: user.password,
-    })
-  ).body.accessToken;
+  tokens = (await request(app).post("/auth/login").send(user)).body;
 });
-
 afterAll(async () => {
-  await Post.deleteMany();
+  await Post.deleteMany({ owner: tokens.userId });
+  await request(app)
+    .get("/auth/logout")
+    .set("Authorization", "jwt " + tokens.refreshToken)
+    .send();
   await mongoose.connection.close();
 });
 
 describe("Post API tests", () => {
-  const postMessage = "this is my test post";
-  let postID: string;
+  const text = "this is my test post";
+  let postId: string;
 
   test("add new post", async () => {
     const response = await request(app)
       .post("/post")
-      .set("Authorization", "jwt " + token)
-      .send({
-        message: postMessage,
-      });
-    postID = response.body._id;
+      .set("Authorization", "jwt " + tokens.accessToken)
+      .send({ text: text });
+    postId = response.body._id;
     expect(response.statusCode).toEqual(200);
     const newPost = response.body;
-    expect(newPost.message).toEqual(postMessage);
+    expect(newPost.text).toEqual(text);
     const response2 = await request(app).get("/post/" + newPost._id);
     expect(response2.statusCode).toEqual(200);
     const post2 = response2.body;
-    expect(post2.message).toEqual(postMessage);
+    expect(post2.text).toEqual(text);
   });
 
   test("get all posts", async () => {
     const response = await request(app).get("/post");
-    const post = response.body[0];
+    const posts = response.body.filter(
+      (element: { text: string; owner: string }) =>
+        (element.owner = tokens.userId)
+    );
     expect(response.statusCode).toEqual(200);
-    expect(response.body.length).toEqual(1);
-    expect(post._id).toEqual(postID);
-    expect(post.message).toEqual(postMessage);
+    expect(posts.length).toEqual(1);
+    expect(posts[0]._id).toEqual(postId);
+    expect(posts[0].text).toEqual(text);
   });
 
   test("get post by id", async () => {
-    const response = await request(app).get(`/post/${postID}`);
+    const response = await request(app).get(`/post/${postId}`);
     const post = response.body;
     expect(response.statusCode).toEqual(200);
-    expect(post._id).toEqual(postID);
-    expect(post.message).toEqual(postMessage);
+    expect(post._id).toEqual(postId);
+    expect(post.text).toEqual(text);
   });
 
   test("get posts by sender", async () => {
@@ -69,32 +65,29 @@ describe("Post API tests", () => {
     const post = response.body[0];
     expect(response.statusCode).toEqual(200);
     expect(response.body.length).toEqual(1);
-    expect(post._id).toEqual(postID);
-    expect(post.message).toEqual(postMessage);
+    expect(post._id).toEqual(postId);
+    expect(post.text).toEqual(text);
   });
 
   test("update post", async () => {
     const response = await request(app)
-      .put(`/post/${postID}`)
-      .set("Authorization", "jwt " + token)
+      .put(`/post/${postId}`)
+      .set("Authorization", "jwt " + tokens.accessToken)
       .send({
-        message: "new message",
+        text: "new message",
       });
     expect(response.statusCode).toEqual(200);
-    const verification_response = await request(app).get(`/post/${postID}`);
+    const verification_response = await request(app).get(`/post/${postId}`);
     const post = verification_response.body;
-    expect(post._id).toEqual(postID);
-    expect(post.message).toEqual("new message");
+    expect(post._id).toEqual(postId);
+    expect(post.text).toEqual("new message");
   });
 
   test("update non-existing post", async () => {
     const response = await request(app)
       .put(`/post/65080d84ca16c0f1f1ba116c`)
-      .set("Authorization", "jwt " + token)
-      .send({
-        message: "new message",
-        sender: "new sender",
-      });
+      .set("Authorization", "jwt " + tokens.accessToken)
+      .send({ text: "new message" });
     expect(response.statusCode).toEqual(404);
   });
 });
